Move API wrapper-related files into /src/utils/api directory
